Modified group_search, verify_group, words.txt, get_users

+++ VkAgent.py
@@ -1,12 +1,9 @@
 import requests
 import os
 import json
-import Ya
 import Agent
 import time
-import random as rnd
 from bs4 import BeautifulSoup
-from pprint import pprint
 import sqlite3 as sq
 
 
@@ -32,16 +29,24 @@ class VkAgent(Agent.Social):
         table_name = suffix if relevant else 'groups_search'
         with sq.connect(os.path.join(self.path_ads, "social_agent.db")) as con:
             cur = con.cursor()
-            cur.execute("PRAGMA FOREIGN_KEYS=ON")
-            cur.execute(f"DROP TABLE IF EXISTS groups_search_users")
-            cur.execute(f"DROP TABLE IF EXISTS users_groups")
-            cur.execute(f"DROP TABLE IF EXISTS {table_name}")
-            cur.execute(f"DROP TABLE IF EXISTS users_list")
+            if not relevant:
+                cur.execute("PRAGMA FOREIGN_KEYS=ON")
+                cur.execute(f"DROP TABLE IF EXISTS groups_search_prev")
+                cur.execute(f"DROP TABLE IF EXISTS groups_search_users")
+                cur.execute(f"DROP TABLE IF EXISTS groups_search")
+                cur.execute(f"DROP TABLE IF EXISTS users_groups")
+                cur.execute(f"DROP TABLE IF EXISTS {table_name}")
+                cur.execute(f"DROP TABLE IF EXISTS users_list")
+
+            cur.execute(f"""CREATE TABLE IF NOT EXISTS groups_search_prev (
+                id INTEGER,                
+                screen_name VARCHAR(50)                                        
+                )""")
 
             cur.execute(f"""CREATE TABLE IF NOT EXISTS {table_name} (
                 id INTEGER PRIMARY KEY,
                 count INTEGER,
-                screen_name TEXT                                        
+                screen_name VARCHAR(50)                                        
                 )""")
 
             cur.execute(f"""CREATE TABLE IF NOT EXISTS users_list (
@@ -49,7 +54,7 @@ class VkAgent(Agent.Social):
                 last_seen_month INTEGER,
                 count_groups INTEGER,
                 city_id INTEGER,
-                sex TEXT,
+                sex VARCHAR(6),
                 stop_list INTEGER
                 )""")
 
@@ -64,13 +69,13 @@ class VkAgent(Agent.Social):
                 group_search_id INTEGER,
                 user_id INTEGER,
                 FOREIGN KEY (group_search_id) REFERENCES groups_search(id),
-                FOREIGN KEY (user_id) REFERENCES users_list(id)
+                FOREIGN KEY (user_id) REFERENCES users_list(id),
+                CONSTRAINT pk PRIMARY KEY (group_search_id, user_id)
                 )""")
 
     def __set_params(self, zero=True):
         self.author = 0 if zero else self.author + 1
         print(f'Токен заменен на >>> {self.author}!')
-        # time.sleep(0.5)
         self.params = {'access_token': self.token[self.author], 'v': '5.131'}
 
     def res_stability(self, method, params_delta, i=0):
@@ -94,7 +99,7 @@ class VkAgent(Agent.Social):
         Условие включения группы в отбор
         """
         with open(os.path.join(os.getcwd(), 'words.txt'), encoding="utf-8") as file:
-            words = [rew.strip().lower() for rew in file.readlines()]
+            words = [rew.strip().lower() for rew in file.readlines()][1:]
         for rew in words:
             if rew == 'stop':
                 flag = False
@@ -133,16 +138,17 @@ class VkAgent(Agent.Social):
             return int(''.join(count.text.split()))
         return -1
 
-    def group_search(self, q: str, members=None, suffix='groups', verify=True, relevant=False):
+    def group_search(self, members=None, suffix='groups', verify=True, relevant=False):
         """
         Поиск групп по ключевой фразе
-        :param q: ключевая фраза для поиска
         :param suffix: суффикс для создаваемого итогового файла
         :param verify: указывает на необходимость проверки по условию verify_group
         :param relevant: указывается True при поиске релевантных групп
         :param members: минимальное количество участников группы
         """
         self.db_create(suffix, relevant)
+        with open(os.path.join(os.getcwd(), 'words.txt'), encoding="utf-8") as file:
+            q = file.readline().strip()
         table_name = suffix if relevant else 'groups_search'
         with sq.connect(os.path.join(self.path_ads, "social_agent.db")) as con:
             cur = con.cursor()
@@ -153,18 +159,28 @@ class VkAgent(Agent.Social):
                                     'offset': offset}
                     response = self.res_stability('groups.search', params_delta)
                     offset += 1
-                    if response and response['response']['items'] and offset < 6:
+                    if response and response['response']['items']:
                         for item in response['response']['items']:
                             print(f"offset={offset}/id{item['id']}")
-                            if (item['id'],) not in cur.execute(f"SELECT id FROM {table_name}").fetchall():
-                                if verify and self.verify_group(item):
-                                    cur.execute(
-                                        f"INSERT INTO {table_name} (id, screen_name) VALUES({item['id']}, '{item['screen_name']}')")
-                                elif not verify:
-                                    cur.execute(
-                                        f"INSERT INTO {table_name} (id, screen_name) VALUES({item['id']}, '{item['screen_name']}')")
+                            if verify and self.verify_group(item):
+                                cur.execute(
+                                    f"INSERT INTO groups_search_prev (id, screen_name) VALUES({item['id']}, '{item['screen_name']}')")
+                            elif not verify:
+                                cur.execute(
+                                    f"INSERT INTO groups_search_prev (id, screen_name) VALUES({item['id']}, '{item['screen_name']}')")
                     else:
                         break
+
+            cur.execute(f"""
+                INSERT INTO {table_name} (id, screen_name)
+                SELECT id, screen_name                  
+                FROM groups_search_prev
+                GROUP BY id, screen_name
+                ORDER BY id
+                """)
+            cur.execute(f"""
+                DELETE FROM groups_search_prev
+                """)
 
             if members:
                 groups = cur.execute(f"SELECT id FROM {table_name}").fetchall()
@@ -172,7 +188,10 @@ class VkAgent(Agent.Social):
                     print(f'+count: id{group[0]}')
                     count = self._get_offset(group)[1]
                     count = count if count != -1 else self.get_count_group(group[0])
-                    cur.execute(f"UPDATE {table_name} SET count = {count} WHERE id = {group[0]}")
+                    if count >= members:
+                        cur.execute(f"UPDATE {table_name} SET count = {count} WHERE id = {group[0]}")
+                    else:
+                        cur.execute(f"DELETE FROM {table_name} WHERE id = {group[0]}")
 
     def _get_offset(self, group_id):
         """
@@ -224,61 +243,54 @@ class VkAgent(Agent.Social):
 
     def get_users(self, sex=1, city=110):
         """
-        Создание списка из id подписчиков групп из файла с id групп
+        Создание списка из id подписчиков групп
         :param sex: пол подписчика (по умолчанию жен)
         :param city: идентификатор города пользователя
-        Результат записывается в файл, готовый к импорту в РК VK
         """
         with sq.connect(os.path.join(self.path_ads, "social_agent.db")) as con:
             cur = con.cursor()
             cur.execute("SELECT * FROM groups_search")
             len_group = len(list(cur))
+            cur.execute("DELETE FROM groups_search_users")
+
+            stop_users = set()
             groups = cur.execute("SELECT id FROM groups_search").fetchall()
             count_i = 1
             for group in groups:
                 print(f'id{group[0]}_{count_i}/{len_group}')
                 users = self.__get_users(group[0], sex=sex, city=city)
+                users = list(set(users) - stop_users)
                 for user in users:
                     cur.execute(f"INSERT INTO groups_search_users VALUES({group[0]}, {user})")
                 count_i += 1
 
-        # считаем количество подписок каждого пользователя на группы из groups_search
-        # with sq.connect(os.path.join(self.path_ads, "social_agent.db")) as con:
-        #     cur = con.cursor()
-        #     cur.execute("SELECT user_id FROM groups_search_users")
-        #     count_groups = {}
-        #     for user in cur:
-        #         count_groups[user[0]] = count_groups.get(user[0], 0) + 1
-
-        # считаем количество подписок каждого пользователя на группы из groups_search
+        # считаем количество подписок каждого пользователя
         with sq.connect(os.path.join(self.path_ads, "social_agent.db")) as con:
             cur = con.cursor()
+            cur.execute("DELETE FROM users_list")
             cur.execute(f"""
-                SELECT user_id, COUNT(user_id) AS count                
+                INSERT INTO users_list (id, last_seen_month, count_groups, city_id, sex, stop_list)
+                SELECT user_id, {self.last_seen}, COUNT(user_id) AS count, {city}, '{'female' if sex == 1 else 'male'}', 0                  
                 FROM groups_search_users
                 GROUP BY user_id
-                ORDER BY count                
+                ORDER BY count
                 """)
-            count_groups = cur.fetchall()
 
-        # записываем полученные данные
+        # записываем данные в файл
         with sq.connect(os.path.join(self.path_ads, "social_agent.db")) as con:
             cur = con.cursor()
-            for user in count_groups:
-                cur.execute(f"""INSERT INTO users_list VALUES(
-                {user[0]},
-                {self.last_seen}, 
-                {user[1]},
-                {city},
-                '{'female' if sex == 1 else 'male'}',
-                0)""")
+            cur.execute(f"""SELECT id FROM users_list
+                            WHERE count_groups >= {self.count_groups}
+                            """)
+            male = 'female' if sex == 1 else 'male'
+            users_file = os.path.join(self.path_users,
+                                      f"{os.path.split(self.path_ads)[1]}_{self.count_groups}_groups_{male}_sex_{city}_city.txt")
+            with open(users_file, 'w', encoding="utf-8") as f:
+                for user in cur:
+                    f.write(f'{user[0]}\n')
 
     def get_user_groups(self, user_id):
-        """
-        Создает кортеж из списка групп пользователя и количество групп
-        :param user_id: id пользователя, для которого создается кортеж
-        :return: обозначенный кортеж из списка и количества групп
-        """
+
         if self._users_lock(user_id):
             return -1, -1
         params_delta = {'user_id': user_id, 'offset': 0}
@@ -300,22 +312,18 @@ class VkAgent(Agent.Social):
         return -1, -1
 
     def get_users_groups(self):
-        """
-        Создает словарь :
-        {'user_id':
-        {'groups': [список из id групп, в которые входит пользователь]
-         'count': количество групп пользователя}
-        ...
-        }
-        Записывает в файлы по 1000 'user_id' в каждом
-        """
-        n = int(input('Пользователь состоит не менеее чем в n групп из поиска: ').strip())
+
         print('Получаем данные по группам пользователей')
         with sq.connect(os.path.join(self.path_ads, "social_agent.db")) as con:
             cur = con.cursor()
             cur1 = con.cursor()
-            all_count = len(list(cur.execute(f"SELECT id FROM users_list WHERE count_groups >= {n}")))
-            cur.execute(f"SELECT id FROM users_list WHERE count_groups >= {n}")
+            all_count = len(list(cur.execute(f"""SELECT id FROM users_list WHERE
+                                                 count_groups >= {self.count_groups}
+                                                 """)))
+            cur.execute("DELETE FROM users_groups")
+            cur.execute(f"""SELECT id FROM users_list WHERE
+                            count_groups >= {self.count_groups}
+                            """)
             for count, user in enumerate(cur, start=1):
                 print(f'{count}/{all_count}_id{user[0]}')
                 user_groups_info = self.get_user_groups(user[0])
@@ -360,15 +368,15 @@ class VkAgent(Agent.Social):
 
 def search_ads():
     def set_groups_param(c):
-        words = input('Введите ключевую фразу для поиска групп: ').strip().lower()
         members = input('Введите минимальное количество членов групп, если нет, то - "N": ').strip().lower()
         if members.isdigit():
-            c.group_search(members=int(members), q=words)
+            c.group_search(members=int(members))
         else:
-            c.group_search(q=words)
+            c.group_search()
 
     folder_name = input(f'Введите название нового проекта либо существующего: ').strip()
     company = VkAgent(folder_name)
+
     print('Выполнить новый поиск групп или использовать ранее выполненный:')
     i = input('"Y" - новый поиcк, "любой символ" - использовать существующий: ').strip().lower()
     if i == 'y':
@@ -377,12 +385,13 @@ def search_ads():
         print('Целевых групп не создано, сначала выполните поиск')
         set_groups_param(c=company)
 
-    if os.listdir(company.path_users):
-        print('Выполнить новый отбор аудитории или использовать ранее выполненный:')
-        i = input('"Y" - новый, "любой символ" - использовать существующий: ').strip().lower()
-        if i == 'y':
-            company.get_users()
-    else:
+    print('Выполнить новый отбор аудитории или использовать существующий:')
+    i = input('"Y" - новый отбор, "любой символ" - использовать существующий: ').strip().lower()
+    if i == 'y':
+        company.get_users()
+    elif not os.path.isfile(os.path.join(company.path_users,
+                                         f"{os.path.split(company.path_ads)[1]}_{company.count_groups}_groups_{male}_sex_{city}_city.txt")):
+        print('Аудитория не отобрана')
         company.get_users()
 
     q = input('Выполнить поиск групп пользователей по всем отобранным пользователям ("Y"/"N"): ').strip().lower()
@@ -392,11 +401,3 @@ def search_ads():
 
 if __name__ == '__main__':
     search_ads()
-    # vk1 = VkAgent('ads_15')
-    # vk1.get_users()
-    # vk1.get_users_groups()
-    # pprint(vk1.friends_info("6055736"))
-
-    # FILE_DIR2 = "Oksa_Studio"
-    # oksa_studio = VkAgent()
-    # oksa_studio.files_downloader('-142029999', FILE_DIR2)
